+++ docs/docs/releases.md
@@ -5,10 +5,13 @@ language: 'en'
 
 # Releases
 
-## Unreleased
+<!-- ## Unreleased -->
 
 <!-- - Fix: MacOS Delete key doesn't work in kitty mode [#513](https://github.com/raphamorim/rio/issues/513). -->
 <!-- - Fix: Kitty keyboard protocol doesn't work with tmux [#599](https://github.com/raphamorim/rio/issues/599). -->
+
+## 0.1.15
+
 - Introduce `cursor.blinking-interval`, default value is 800ms.
 - Fix blinking cursor lag issue.
 - Fix adaptive theme background color on macos.